django_carsharing/requirements.txt
drf_spectacular==0.26.5
django-allauth==0.58.0
django-clearcache==1.2.1
amqp==5.1.1
asgiref==3.6.0
astroid==2.15.5
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.6
Django==4.2.1
django-clearcache==1.2.1
django-cors-headers==4.0.0
django-filter==23.2
django-oauth-toolkit==2.3.0
django-redis==5.3.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-social-oauth2==2.1.3
drf-yasg==1.21.6
ecdsa==0.18.0
flower==2.0.0
gunicorn==20.1.0
humanize==4.6.0
idna==3.4
inflection==0.5.1
isort==5.12.0
jwcrypto==1.5.0
kombu==5.3.1
lazy-object-proxy==1.9.0
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
platformdirs==3.5.1
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.7.0
pylint==2.17.4
python-dateutil==2.8.2
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
tornado==6.3.2
typing_extensions==4.6.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
